Refactor create shared album
